backend/requirements.txt
Flask==2.1.0
Flask-Cors==3.0.10
spacy==3.5.0
scikit-learn==1.0.2
pandas==1.4.2
google-generativeai==0.3.1
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.33.0
python-dotenv==0.19.2
requests==2.27.1
numpy==1.22.3
pytest==7.1.1
gunicorn==20.1.0
pyautogen==0.1.0